refactor: repository에 flush() 메서드 추가
